+++ src/spec/components/App.spec.tsx
@@ -2,6 +2,10 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import App from '@components/App';
 
+afterEach(() => {
+  jest.restoreAllMocks();
+});
+
 describe('App', () => {
   it('should render correctly', () => {
     const tree = renderer